# server-go/docs/swagger.yaml
basePath: /api
...
host: localhost:3001
info:
  contact:
    email: support@example.com
    name: Dashboard Team
  description: Dashboard system with AI-powered chart generation
  title: TRVE Dashboards API
  version: "1.0"
paths:
...
  /api/frigate/{connection_id}/cameras:
    get:
      description: Returns camera names from a Frigate NVR connection
      parameters:
      - description: Connection ID
        in: path
        name: connection_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Frigate camera list
      tags:
      - Frigate
  /api/frigate/{connection_id}/event/{event_id}/clip:
    get:
      description: Returns the MP4 clip for a detection event (supports Range headers)
      parameters:
      - description: Connection ID
        in: path
        name: connection_id
        required: true
        type: string
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - video/mp4
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Get event clip
      tags:
      - Frigate
  /api/frigate/{connection_id}/event/{event_id}/snapshot:
    get:
      description: Returns the JPEG snapshot for a detection event
      parameters:
      - description: Connection ID
        in: path
        name: connection_id
        required: true
        type: string
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Get event snapshot
      tags:
      - Frigate
  /api/frigate/{connection_id}/events/{camera}:
    get:
      description: Returns recent detection events for a camera
      parameters:
      - description: Connection ID
        in: path
        name: connection_id
        required: true
        type: string
      - description: Camera name
        in: path
        name: camera
        required: true
        type: string
      - default: 10
        description: Max events to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Get camera events
      tags:
      - Frigate
  /api/frigate/{connection_id}/info:
    get:
      description: Returns connection metadata needed for direct WebSocket access
        (go2rtc MSE)
      parameters:
      - description: Connection ID
        in: path
        name: connection_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Frigate connection info
      tags:
      - Frigate
...
  /api/frigate/{connection_id}/snapshot/{camera}:
    get:
      description: Returns the latest JPEG snapshot for a camera
      parameters:
      - description: Connection ID
        in: path
        name: connection_id
        required: true
        type: string
      - description: Camera name
        in: path
        name: camera
        required: true
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Get camera snapshot
      tags:
      - Frigate
...
swagger: "2.0"